Move DeepSSMParameters to Project library
