+++ Studio/CMakeLists.txt
@@ -157,7 +157,6 @@ SET(STUDIO_ANALYSIS_MOC_HDRS
 
 SET(STUDIO_DEEPSSM_SRCS
   DeepSSM/DeepSSMTool.cpp
-  DeepSSM/DeepSSMParameters.cpp
   DeepSSM/DeepSSMJob.cpp
   )
 SET(STUDIO_DEEPSSM_MOC_HDRS
