+++ Studio/DeepSSM/DeepSSMTool.cpp
@@ -16,7 +16,7 @@
 #include <Data/ShapeWorksWorker.h>
 #include <Data/Worker.h>
 #include <DeepSSM/DeepSSMJob.h>
-#include <DeepSSM/DeepSSMParameters.h>
+#include <Project/DeepSSMParameters.h>
 #include <DeepSSM/DeepSSMTool.h>
 #include <Interface/ShapeWorksStudioApp.h>
 #include <Interface/Style.h>
